requirements.txt
pdfplumber==0.10.2
langchain>=0.1.0
openai>=1.0.0
pymilvus==2.3.4
python-docx==1.1.0
rank_bm25==0.2.2
scikit-learn>=1.3.0
dotenv
langchain_groq
langchain_huggingface
marshmallow==3.19.0
streamlit
argparse
streamlit-chat
torch==2.1.2
langchain_experimental
numpy==1.26.4
